refactor: poc
